Backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
SQLAlchemy==2.0.37
asyncpg==0.29.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
greenlet==3.1.1
openai==1.61.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
twilio==9.2.2
python-multipart==0.0.9
pgvector==0.3.6
tiktoken==0.8.0
cryptography==42.0.5
PyJWT==2.9.0
